cache tests using mvc integration framework
