+++ Src/Integration/Tests/Cache/Constants.cs
@@ -1,0 +1,8 @@
+namespace Tests.Cache
+{
+    public static class Constants
+    {
+        public static readonly string WasCachedResult = "Was Cached:True";
+        public static readonly string NotCachedResult = "Was Cached:False";
+    }
+}
